requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
Django==3.0.3
gast==0.3.3
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pytz==2019.3
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==1.0.0
wrapt==1.11.2